[Travis] Adding support for forks testing
This CL allows to add testing support in Travis for forks leveraging
the `dep ensure --add` option to support this use case.

As of today, `dep ensure` will pull the official prebid-server repo and
compile it but won't rely on the forked repo.

This change needs to stay in the fork to ensure Travis multi-package builds work.